+++ src/main/java/com/ams/backend/repository/AnswerRepository.java
@@ -3,6 +3,6 @@ package com.ams.backend.repository;
 import com.ams.backend.entity.Answer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
 
 }
