+++ src/main/java/com/ams/backend/repository/AuditTypeRepository.java
@@ -3,6 +3,6 @@ package com.ams.backend.repository;
 import com.ams.backend.entity.AuditType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditTypeRepository extends JpaRepository<AuditType, Long> {
+public interface AuditTypeRepository extends JpaRepository<AuditType, Integer> {
 
 }
